+++ src/UrlHelperFactory.php
@@ -16,6 +16,34 @@ use function sprintf;
 
 class UrlHelperFactory
 {
+    /** @var string Base path for the URL helper */
+    private $basePath;
+
+    /** @var string $routerServiceName */
+    private $routerServiceName;
+
+    /**
+     * Allow serialization
+     */
+    public static function __set_state(array $data) : self
+    {
+        return new self(
+            $data['basePath'] ?? '/',
+            $data['routerServiceName'] ?? RouterInterface::class
+        );
+    }
+
+    /**
+     * Allows varying behavior per-instance.
+     *
+     * Defaults to '/' for the base path, and the FQCN of the RouterInterface.
+     */
+    public function __construct(string $basePath = '/', string $routerServiceName = RouterInterface::class)
+    {
+        $this->basePath = $basePath;
+        $this->routerServiceName = $routerServiceName;
+    }
+
     /**
      * Create a UrlHelper instance.
      *
@@ -23,14 +51,16 @@ class UrlHelperFactory
      */
     public function __invoke(ContainerInterface $container) : UrlHelper
     {
-        if (! $container->has(RouterInterface::class)) {
+        if (! $container->has($this->routerServiceName)) {
             throw new Exception\MissingRouterException(sprintf(
                 '%s requires a %s implementation; none found in container',
                 UrlHelper::class,
-                RouterInterface::class
+                $this->routerServiceName
             ));
         }
 
-        return new UrlHelper($container->get(RouterInterface::class));
+        $helper = new UrlHelper($container->get($this->routerServiceName));
+        $helper->setBasePath($this->basePath);
+        return $helper;
     }
 }